indices now unsigned int, addded model importer for tests

+++ TD3D/TDXBufferManager.h
@@ -10,25 +10,25 @@ class TDXBufferManager
 public:
 	~TDXBufferManager();
 
-	void CreateStaticVertexBuffer(std::string name, void* vertices, unsigned short numVertices, unsigned short vertexSize, ID3D11Device* device);
+	void CreateStaticVertexBuffer(std::string name, void* vertices, unsigned int numVertices, unsigned short vertexSize, ID3D11Device* device);
 
-	void CreateStaticIndexBuffer(std::string name, unsigned short* indices, unsigned short numIndices, ID3D11Device* device);
+	void CreateStaticIndexBuffer(std::string name, unsigned int* indices, unsigned int numIndices, ID3D11Device* device);
 
-	void CreateDynamicConstantBuffer(std::string name, unsigned short structSize, ID3D11Device* device);
+	void CreateDynamicConstantBuffer(std::string name, unsigned int structSize, ID3D11Device* device);
 
-	void CreateDynamicVertexBuffer(std::string name, void* vertices, unsigned short numVertices, unsigned short vertexSize, ID3D11Device* device);
+	void CreateDynamicVertexBuffer(std::string name, void* vertices, unsigned int numVertices, unsigned short vertexSize, ID3D11Device* device);
 
-	void CreateDynamicIndexBuffer(std::string name, unsigned short* indices, unsigned short numIndices, ID3D11Device* device);
+	void CreateDynamicIndexBuffer(std::string name, unsigned int* indices, unsigned int numIndices, ID3D11Device* device);
 
 	void UpdateStaticVertexBuffer(std::string name, void* vertices, ID3D11DeviceContext* context);
 
-	void UpdateStaticIndexBuffer(std::string name, unsigned short* indices, ID3D11DeviceContext* context);
+	void UpdateStaticIndexBuffer(std::string name, unsigned int* indices, ID3D11DeviceContext* context);
 
 	void UpdateDynamicConstantBuffer(std::string name, void* constantBufferStruct, unsigned short structSize, ID3D11DeviceContext* context);
 
-	void UpdateDynamicVertexBuffer(std::string name, void* vertices, unsigned short numVertices, unsigned short vertexSize, ID3D11DeviceContext* context);
+	void UpdateDynamicVertexBuffer(std::string name, void* vertices, unsigned int numVertices, unsigned short vertexSize, ID3D11DeviceContext* context);
 
-	void UpdateDynamicIndexBuffer(std::string name, unsigned short* indices, unsigned short numIndices, ID3D11DeviceContext* context);
+	void UpdateDynamicIndexBuffer(std::string name, unsigned int* indices, unsigned int numIndices, ID3D11DeviceContext* context);
 
 	void BindVertexBuffer(std::string vertexName, UINT stride, UINT offset, ID3D11DeviceContext* context);
 	void BindIndexBuffer(std::string indexName, ID3D11DeviceContext* context);
