+++ TD3D/TDXRenderDevice.h
@@ -24,8 +24,8 @@ public:
 	virtual void BeginFrame(float r, float g, float b, float a) override;
 	virtual void EndFrame() override;
 
-	virtual void Draw(unsigned short numIndices, unsigned short startIndexLocation, unsigned short baseVertexLocation) override;
-	virtual void DrawSprite(unsigned short numVertices, unsigned short startVertexLocation) override;
+	virtual void Draw(unsigned int numIndices, unsigned int startIndexLocation, unsigned int baseVertexLocation) override;
+	virtual void DrawSprite(unsigned int numVertices, unsigned int startVertexLocation) override;
 	virtual void RenderText(const wchar_t* text, float x, float y) override;
 
 	virtual void SetProjectionValues(float fovDegrees, float aspectRatio, float nearZ, float farZ) override;
@@ -38,15 +38,15 @@ public:
 	virtual void BindVertexShader(std::string name) override;
 	virtual void BindPixelShader(std::string name) override;
 
-	virtual void CreateStaticVertexBuffer(std::string name, void* vertices, unsigned short numVertices, unsigned short vertexSize) override;
-	virtual void CreateStaticIndexBuffer(std::string name, unsigned short* indices, unsigned short numIndices) override;
-	virtual void CreateDynamicVertexBuffer(std::string name, void* vertices, unsigned short numVertices, unsigned short vertexSize) override;
-	virtual void CreateDynamicIndexBuffer(std::string name, unsigned short* indices, unsigned short numIndices) override;
+	virtual void CreateStaticVertexBuffer(std::string name, void* vertices, unsigned int numVertices, unsigned short vertexSize) override;
+	virtual void CreateStaticIndexBuffer(std::string name, unsigned int* indices, unsigned int numIndices) override;
+	virtual void CreateDynamicVertexBuffer(std::string name, void* vertices, unsigned int numVertices, unsigned short vertexSize) override;
+	virtual void CreateDynamicIndexBuffer(std::string name, unsigned int* indices, unsigned int numIndices) override;
 
 	virtual void UpdateStaticVertexBuffer(std::string name, void* vertices) override;
-	virtual void UpdateStaticIndexBuffer(std::string name, unsigned short* indices) override;
-	virtual void UpdateDynamicVertexBuffer(std::string name, void* vertices, unsigned short numVertices, unsigned short vertexSize) override;
-	virtual void UpdateDynamicIndexBuffer(std::string name, unsigned short* indices, unsigned short numIndices) override;
+	virtual void UpdateStaticIndexBuffer(std::string name, unsigned int* indices) override;
+	virtual void UpdateDynamicVertexBuffer(std::string name, void* vertices, unsigned int numVertices, unsigned short vertexSize) override;
+	virtual void UpdateDynamicIndexBuffer(std::string name, unsigned int* indices, unsigned int numIndices) override;
 
 	virtual void BindVertexBuffer(std::string vertexName, UINT stride, UINT offset) override;
 	virtual void BindIndexBuffer(std::string indexName) override;
